chore: Remove single update task and replace with multiple
