+++ src/index.ts
@@ -15,7 +15,7 @@ import { tasksListCompleted } from './tools/tasks-list-completed.js'
 import { tasksListForContainer } from './tools/tasks-list-for-container.js'
 import { tasksOrganizeMultiple } from './tools/tasks-organize-multiple.js'
 import { tasksSearch } from './tools/tasks-search.js'
-import { tasksUpdateOne } from './tools/tasks-update-one.js'
+import { tasksUpdateMultiple } from './tools/tasks-update-multiple.js'
 
 const tools = {
     projectsList,
@@ -29,7 +29,7 @@ const tools = {
     tasksCompleteMultiple,
     tasksSearch,
     tasksAddMultiple,
-    tasksUpdateOne,
+    tasksUpdateMultiple,
     tasksOrganizeMultiple,
     overview,
 }
@@ -48,7 +48,7 @@ export {
     tasksCompleteMultiple,
     tasksSearch,
     tasksAddMultiple,
-    tasksUpdateOne,
+    tasksUpdateMultiple,
     tasksOrganizeMultiple,
     overview,
 }
